refactor(orderbook): move table component to own file
